requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
asgiref==3.11.0
sqlparse==0.5.4
